refactor: split SDL core runtime-services modules into packages
Convert composition, orchestration, participant_behavior, and
runtime_mail_service into same-named packages behind API-stable
re-export facades (issue #52). Pure refactor: public identities,
Pydantic/enum metadata, validator order, error wording, insertion
order, and generated schemas are all preserved; no external import
line changes.

Drain the four oversized_allowlist.yaml entries (allowlist now empty)
and retarget every governance/doc path reference to the package
facades (SDL lineage ledger + its identity-cutover digest pin,
scientific-completeness assessment, catalog parity + references.md,
shared-semantic-integrity note, sdl.rst autodoc). The frozen
_ADR015_INITIAL_OVERSIZED_FILES baseline is left untouched.
