+++ contracts/provenance/sdl-lineage-ledger-v1.json
@@ -1765,7 +1765,7 @@
           ],
           "aces_boundaries": [
             {
-              "artifact": "implementations/python/packages/raes/orchestration.py",
+              "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
               "symbol_or_pointer": "events"
             }
           ],
@@ -2050,7 +2050,7 @@
           ],
           "aces_boundaries": [
             {
-              "artifact": "implementations/python/packages/raes/orchestration.py",
+              "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
               "symbol_or_pointer": "injects"
             }
           ],
@@ -2504,7 +2504,7 @@
           ],
           "aces_boundaries": [
             {
-              "artifact": "implementations/python/packages/raes/orchestration.py",
+              "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
               "symbol_or_pointer": "scripts"
             }
           ],
@@ -2568,7 +2568,7 @@
           ],
           "aces_boundaries": [
             {
-              "artifact": "implementations/python/packages/raes/orchestration.py",
+              "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
               "symbol_or_pointer": "stories"
             }
           ],
@@ -4395,7 +4395,7 @@
           ],
           "aces_boundaries": [
             {
-              "artifact": "implementations/python/packages/raes/orchestration.py",
+              "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
               "symbol_or_pointer": "events"
             }
           ],
@@ -4893,7 +4893,7 @@
           "symbol_or_pointer": "Entity model lineage"
         },
         {
-          "artifact": "implementations/python/packages/raes/orchestration.py",
+          "artifact": "implementations/python/packages/raes/orchestration/__init__.py",
           "symbol_or_pointer": "Inject, Event, Script, and Story model lineage"
         },
         {
