+++ specs/formal/participant-semantics/README.md
@@ -779,7 +779,7 @@ Minimum future implementation artifacts:
 
 Current implementation artifacts for the `SEM-208` slice:
 
-- `implementations/python/packages/raes/participant_behavior.py` defines
+- `implementations/python/packages/raes/participant_behavior/__init__.py` defines
   typed action contracts and observation boundaries;
 - `implementations/python/packages/raes/semantics/participant_behavior.py`
   and `implementations/python/packages/raes/validator/` fail closed on
@@ -821,7 +821,7 @@ Minimum future implementation artifacts:
 
 Current implementation artifacts for the `SEM-209` slice:
 
-- `implementations/python/packages/raes/participant_behavior.py` defines
+- `implementations/python/packages/raes/participant_behavior/__init__.py` defines
   interaction classes, target references, related actions, and shared-state
   references on action contracts;
 - `implementations/python/packages/raes/semantics/participant_behavior.py`
@@ -890,7 +890,7 @@ Implemented transition discipline:
 
 Current implementation artifacts for the `SEM-210` slice:
 
-- `implementations/python/packages/raes/participant_behavior.py` defines
+- `implementations/python/packages/raes/participant_behavior/__init__.py` defines
   participant information-boundary classes, view dispositions, explicit view
   rules, time-indexed view transitions, realized-view disclosure metadata, and
   observation-boundary hidden, observable, and evidence-only reference
@@ -971,7 +971,7 @@ Current implementation artifacts for the `SEM-211` slice:
 - `implementations/python/packages/raes/participant_action_semantics.py`
   defines controlled precondition, effect, and portable failure vocabularies
   plus typed action-contract declarations and backend failure mappings;
-- `implementations/python/packages/raes/participant_behavior.py` embeds
+- `implementations/python/packages/raes/participant_behavior/__init__.py` embeds
   those typed declarations in governed participant action contracts;
 - `implementations/python/packages/raes_processor/compiler/` carries the
   typed precondition classes, effect classes, failure classes, and backend
@@ -1066,7 +1066,7 @@ Current implementation artifacts for the first `SEM-213` slice:
   defines typed time domains, temporal event points, schedule/cadence/deadline/
   dwell/latency/time-window contract kinds, backend timing disclosure kinds,
   support modes, and abstract temporal states;
-- `implementations/python/packages/raes/participant_behavior.py` embeds
+- `implementations/python/packages/raes/participant_behavior/__init__.py` embeds
   temporal contracts and backend timing disclosures in governed participant
   action contracts, requires temporal preconditions to resolve to typed
   temporal contracts, and fails closed on unknown backend disclosure refs;
